finish EmotionDiaryController and EmotionDiaryServiceImpl backend

+++ EmoSync-backend/emosync-pojo/src/main/java/com/emosync/DTO/query/EmotionDiaryQueryDTO.java
@@ -52,10 +52,10 @@ public class EmotionDiaryQueryDTO {
 
     @Schema(description = "页码", example = "1")
     @Min(value = 1, message = "页码不能小于1")
-    private Long current = 1L;
+    private Integer current = 1;
 
     @Schema(description = "每页大小", example = "10")
     @Min(value = 1, message = "每页大小不能小于1")
     @Max(value = 100, message = "每页大小不能大于100")
-    private Long size = 10L;
+    private Integer size = 10;
 }
